fix: build only from Go version 1.7 onward
